# Implement NarrowUtf16ToAscii for AArch64