--- a/src/libraries/System.Private.CoreLib/src/System/Text/ASCIIUtility.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Text/ASCIIUtility.cs
@@ -1433,9 +1433,8 @@ namespace System.Text
             {
                 if (Sse41.IsSupported)
                 {
-                    Vector128<ushort> asciiMaskForTestZ = Vector128.Create((ushort)0xFF80); // used for PTEST on supported hardware
-                    // If a non-ASCII bit is set in any WORD of the combined vector, we have seen non-ASCII data.
-                    // Jump to the non-ASCII handler to figure out which particular vector contained non-ASCII data.
+                    Vector128<ushort> asciiMaskForTestZ = Vector128.Create((ushort)0xFF80);
+                    // If a non-ASCII bit is set in any WORD of the vector, we have seen non-ASCII data.
                     if (!Sse41.TestZ(utf16Vector.AsInt16(), asciiMaskForTestZ.AsInt16()))
                     {
                         return true;
@@ -1443,7 +1442,7 @@ namespace System.Text
                 }
                 else
                 {
-                    Vector128<ushort> asciiMaskForAddSaturate = Vector128.Create((ushort)0x7F80); // used for PADDUSW
+                    Vector128<ushort> asciiMaskForAddSaturate = Vector128.Create((ushort)0x7F80);
                     // The operation below forces the 0x8000 bit of each WORD to be set iff the WORD element
                     // has value >= 0x0800 (non-ASCII). Then we'll treat the vector as a BYTE vector in order
                     // to extract the mask. Reminder: the 0x0080 bit of each WORD should be ignored.
@@ -1455,10 +1454,12 @@ namespace System.Text
             }
             else if (AdvSimd.Arm64.IsSupported)
             {
+                // First we pick four chars, a larger one from all four pairs of adjecent chars in the vector.
+                // If any of those four chars has a non-ASCII bit set, we have seen non-ASCII data.
                 Vector128<ushort> maxChars = AdvSimd.Arm64.MaxPairwise(utf16Vector, utf16Vector);
                 if ((maxChars.AsUInt64().ToScalar() & 0xFF80FF80FF80FF80) != 0)
                 {
-                    return true; // found non-ASCII data
+                    return true;
                 }
             }
             else
@@ -1593,7 +1594,7 @@ namespace System.Text
             // See comments in GetIndexOfFirstNonAsciiChar_Sse2 for information about how this works.
             if (VectorContainsNonAsciiChar(utf16VectorFirst))
             {
-                goto Finish; // found non-ASCII data
+                goto Finish;
             }
 
             // First part was all ASCII, narrow and aligned write. Note we're only filling in the low half of the vector.