Fixes #41292  partially.