@SwapnilGaikwad 2022-06-01
> Thanks @SwapnilGaikwad for your contribution. Could you also share some performance numbers? You can start with https://github.com/dotnet/performance/blob/d7dac8a7ca12a28d099192f8a901cf8e30361384/src/benchmarks/micro/libraries/System.Text.Encoding/Perf.Encoding.cs.

Hi Kunal, did a quick test on A72 for the `GetBytes` method from the `System.Text.Tests.Perf_Encoding` class. The patch executes about 6% (for ascii) and 3% (for utf-8) strings of 2048 size relative to the main.
However, had to comment out the debug asserts to avoid emitting them in assembly. We are inspecting the assembly further to spot suboptimal sequence of instructions ('ll work on this in the next week due to public holidays).