@SwapnilGaikwad 2022-06-15
Hi @kunalspathak, now the patch makes use of the vector API more. For ASCII strings of 512 size, it executes in about 0.92x on AArch64 and 0.86x on x86 compared to the execution time for the HEAD.
The generic vector implementation (in the Vector.IsHardwareAccelerated block) performs chunked reads so the improvement is not as significant as one would expect while moving from scalar to SIMD version.
Do you have any recommendations to extract the assembly for the intrinsic? Couldn't get it using `--disassm` option from [the docs](https://github.com/dotnet/performance/blob/d7dac8a7ca12a28d099192f8a901cf8e30361384/docs/benchmarkdotnet.md#Disassembly).