@SwapnilGaikwad 2022-06-16
> > Hi @kunalspathak, here are the numbers.
> 
> Thanks for sharing the numbers. It seems the ascii-16 bytes is slightly slower. Do you know why?

It seems the loop peeling logic to get the destination pointing to 8-byte aligned address adds a slight overhead for smaller strings.
However, for longer string the loop peeling logic pays off. If we remove the 8-byte aligned write logic, we see very negligible overhead compared to the current HEAD [1].

[1] **Microbenchmark results** after removing the aligned write logic:
release_runtime_intrinsic: The current PR
release_without_peel: The current PR without aligned write (commenting lines [1551](https://github.com/dotnet/runtime/pull/70080/files#diff-66bbe89271f826c9232bd146abb678844754515dc027f70ad0ce36f751da46ebR1551) to [1570](https://github.com/dotnet/runtime/pull/70080/files#diff-66bbe89271f826c9232bd146abb678844754515dc027f70ad0ce36f751da46ebR1570))
runtime_HEAD: Current HEAD

```
|   Method |        Job |                                                                                                     Toolchain | size | encName |      Mean |    Error |   StdDev |    Median |       Min |       Max | Ratio | MannWhitney(2%) |  Gen 0 | Allocated | Alloc Ratio |
|--------- |----------- |-------------------------------------------------------------------------------------------------------------- |----- |-------- |----------:|---------:|---------:|----------:|----------:|----------:|------:|---------------- |-------:|----------:|------------:|
| GetBytes | Job-TLYBOZ | /release_runtime_intrinsic/bin/testhost/net7.0-Linux-Release-arm64/shared/Microsoft.NETCore.App/7.0.0/corerun |   16 |   ascii |  34.72 ns | 0.039 ns | 0.037 ns |  34.72 ns |  34.66 ns |  34.78 ns |  1.07 |          Slower | 0.0765 |      40 B |        1.00 |
| GetBytes | Job-NLAVLI |      /release_without_peel/bin/testhost/net7.0-Linux-Release-arm64/shared/Microsoft.NETCore.App/7.0.0/corerun |   16 |   ascii |  32.07 ns | 0.058 ns | 0.051 ns |  32.06 ns |  31.95 ns |  32.15 ns |  0.99 |            Same | 0.0764 |      40 B |        1.00 |
| GetBytes | Job-AMSBUC |              /runtime_HEAD/bin/testhost/net7.0-Linux-Release-arm64/shared/Microsoft.NETCore.App/7.0.0/corerun |   16 |   ascii |  32.48 ns | 0.046 ns | 0.043 ns |  32.49 ns |  32.41 ns |  32.54 ns |  1.00 |            Base | 0.0765 |      40 B |        1.00 |
|          |            |                                                                                                               |      |         |           |          |          |           |           |           |       |                 |        |           |             |
| GetBytes | Job-YCMYKR | /release_runtime_intrinsic/bin/testhost/net7.0-Linux-Release-arm64/shared/Microsoft.NETCore.App/7.0.0/corerun |  512 |   ascii | 158.93 ns | 0.265 ns | 0.221 ns | 158.93 ns | 158.52 ns | 159.43 ns |  0.90 |          Faster | 1.0241 |     536 B |        1.00 |
| GetBytes | Job-NLAVLI |      /release_without_peel/bin/testhost/net7.0-Linux-Release-arm64/shared/Microsoft.NETCore.App/7.0.0/corerun |  512 |   ascii | 168.20 ns | 1.472 ns | 1.377 ns | 168.33 ns | 166.01 ns | 170.06 ns |  0.96 |          Faster | 1.0246 |     536 B |        1.00 |
| GetBytes | Job-AMSBUC |              /runtime_HEAD/bin/testhost/net7.0-Linux-Release-arm64/shared/Microsoft.NETCore.App/7.0.0/corerun |  512 |   ascii | 175.96 ns | 0.719 ns | 0.672 ns | 175.88 ns | 174.48 ns | 177.05 ns |  1.00 |            Base | 1.0239 |     536 B |        1.00 |
```